api/requirements.txt
Flask
google-generativeai
python-dotenv
PyPDF2
requests